+++ src/ER_OLEDM1_CH1115_graphics.cpp
@@ -586,22 +586,22 @@ CH1115_Return_Codes_e ERMCH1115_graphics::drawChar(int16_t x, int16_t y, unsigne
     {
            	switch (_FontNumber) {
 #ifdef CH1115_Font_One
-				case CH1115Font_Default : line = pFontDefaultptr[((character - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+				case CH1115Font_Default : line = pgm_read_byte(pFontDefaultptr + ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif 
 #ifdef CH1115_Font_Two
-				case CH1115Font_Thick : line = pFontThickptr[((character  - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+						case CH1115Font_Thick : line = pgm_read_byte(pFontThickptr+ ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif
 #ifdef CH1115_Font_Three
-				case CH1115Font_Seven_Seg : line = pFontSevenSegptr[((character  - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+						case CH1115Font_Seven_Seg: line = pgm_read_byte(pFontSevenSegptr + ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif
 #ifdef CH1115_Font_Four
-				case CH1115Font_Wide: line = pFontWideptr[((character  - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+						case CH1115Font_Wide : line = pgm_read_byte(pFontWideptr + ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif
 #ifdef CH1115_Font_Five
-				case CH1115Font_Tiny : line = pFontTinyptr[((character  - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+						case CH1115Font_Tiny : line = pgm_read_byte(pFontTinyptr + ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif
 #ifdef CH1115_Font_Six
-				case CH1115Font_Homespun: line = pFontHomeSpunptr[((character  - _CurrentFontoffset) * _CurrentFontWidth) + i]; break;
+						case CH1115Font_Homespun : line = pgm_read_byte(pFontHomeSpunptr + ((character - _CurrentFontoffset) * _CurrentFontWidth) + i); break;
 #endif
 				default: // wrong font number
 					return CH1115_WrongFont;
@@ -866,22 +866,28 @@ CH1115_Return_Codes_e ERMCH1115_graphics::drawChar(uint8_t x, uint8_t y, uint8_t
 		switch (_FontNumber)
 		{
 #ifdef CH1115_Font_Seven
-			case CH1115Font_Bignum: ctemp = pFontBigNum16x32ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_Bignum: 
+				ctemp = pgm_read_byte(&pFontBigNum16x32ptr[character - _CurrentFontoffset][i]);break;
 #endif
 #ifdef CH1115_Font_Eight
-			case CH1115Font_Mednum: ctemp = pFontMedNum16x16ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_Mednum: 
+				ctemp = pgm_read_byte(&pFontMedNum16x16ptr[character - _CurrentFontoffset][i]); break;
 #endif
 #ifdef CH1115_Font_Nine
-			case CH1115Font_ArialRound: ctemp = pFontArial16x24ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_ArialRound: 
+				ctemp = pgm_read_byte(&pFontArial16x24ptr[character - _CurrentFontoffset][i]);break;
 #endif
 #ifdef CH1115_Font_Ten
-			case CH1115Font_ArialBold: ctemp = pFontArial16x16ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_ArialBold:
+				ctemp = pgm_read_byte(&pFontArial16x16ptr[character - _CurrentFontoffset][i]);break;
 #endif
 #ifdef CH1115_Font_Eleven
-			case CH1115Font_Mia: ctemp = pFontMia8x16ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_Mia: 
+			ctemp = pgm_read_byte(&pFontMia8x16ptr[character - _CurrentFontoffset][i]); break;
 #endif
 #ifdef CH1115_Font_Twelve
-			case CH1115Font_Dedica: ctemp = pFontDedica8x12ptr[character - _CurrentFontoffset][i]; break;
+			case CH1115Font_Dedica: 
+			ctemp = pgm_read_byte(&pFontDedica8x12ptr[character - _CurrentFontoffset][i]); break;
 #endif
 			default :
 				return CH1115_WrongFont;